Move manage reservations  page into its own folder
