+++ src/domain/reservation/_reservation.scss
@@ -1,3 +1,3 @@
-@import 'manage/manageReservationsPage';
+@import 'manage/page/manageReservationsPage';
 @import 'manage/list/manageReservationsList';
 @import 'manage/status/manageReservationsStatus';
